+++ packages/gatsby/src/schema/__tests__/infer-graphql-type-test.js
@@ -8,7 +8,10 @@ const path = require(`path`)
 const normalizePath = require(`normalize-path`)
 const { clearTypeExampleValues } = require(`../data-tree-utils`)
 const { typeConflictReporter } = require(`../type-conflict-reporter`)
-const { inferObjectStructureFromNodes, clearUnionTypes } = require(`../infer-graphql-type`)
+const {
+  inferObjectStructureFromNodes,
+  clearUnionTypes,
+} = require(`../infer-graphql-type`)
 const { clearTypeNames } = require(`../create-type-name`)
 
 function queryResult(nodes, fragment, { types = [], ignoreFields } = {}) {
@@ -84,6 +87,9 @@ describe(`GraphQL type inferance`, () => {
       "with-hyphen": 2,
       "with resolver": `1012-11-01`,
       123: 42,
+      456: {
+        testingTypeNameCreation: true,
+      },
       aBoolean: true,
       externalUrl: `https://example.com/awesome.jpg`,
       domain: `pizza.com`,
@@ -239,6 +245,9 @@ describe(`GraphQL type inferance`, () => {
         with_hyphen
         with_resolver(formatString:"DD.MM.YYYY")
         _123
+        _456 {
+          testingTypeNameCreation
+        }
       `
     )
 
@@ -251,6 +260,7 @@ describe(`GraphQL type inferance`, () => {
     expect(result.data.listNode[0].with_resolver).toEqual(`01.11.1012`)
     expect(result.data.listNode[0]._123).toEqual(42)
     expect(result.data.listNode[1]._123).toEqual(24)
+    expect(result.data.listNode[0]._456).toEqual(nodes[0][`456`])
   })
 
   describe(`Handles dates`, () => {
@@ -734,7 +744,7 @@ describe(`GraphQL type inferance`, () => {
 
       it(`Uses same union type for same child node types and key`, () => {
         const fields = inferObjectStructureFromNodes({
-          nodes: [{ test___NODE: [`pet_1`, `child_1`] } ],
+          nodes: [{ test___NODE: [`pet_1`, `child_1`] }],
           types,
         })
         const fields2 = inferObjectStructureFromNodes({
@@ -743,7 +753,6 @@ describe(`GraphQL type inferance`, () => {
         })
         expect(fields.test.type).toEqual(fields2.test.type)
       })
-
 
       it(`Uses a different type for the same child node types with a different key`, () => {
         const fields = inferObjectStructureFromNodes({
@@ -763,7 +772,7 @@ describe(`GraphQL type inferance`, () => {
           payload: { id: `toy_1`, internal: { type: `Toy` } },
         })
         const fields = inferObjectStructureFromNodes({
-          nodes: [{ test___NODE: [`pet_1`, `child_1`] } ],
+          nodes: [{ test___NODE: [`pet_1`, `child_1`] }],
           types,
         })
         const fields2 = inferObjectStructureFromNodes({
@@ -774,7 +783,7 @@ describe(`GraphQL type inferance`, () => {
       })
 
       it(`Creates a new type after schema updates clear union types`, () => {
-        const nodes = [{ test___NODE: [`pet_1`, `child_1`] } ]
+        const nodes = [{ test___NODE: [`pet_1`, `child_1`] }]
         const fields = inferObjectStructureFromNodes({ nodes, types })
         clearUnionTypes()
         const updatedFields = inferObjectStructureFromNodes({ nodes, types })
@@ -782,7 +791,7 @@ describe(`GraphQL type inferance`, () => {
       })
 
       it(`Uses a reliable naming convention`, () => {
-        const nodes = [{ test___NODE: [`pet_1`, `child_1`] } ]
+        const nodes = [{ test___NODE: [`pet_1`, `child_1`] }]
         inferObjectStructureFromNodes({ nodes, types })
         clearUnionTypes()
         const updatedFields = inferObjectStructureFromNodes({ nodes, types })