+++ www/src/templates/template-blog-post.js
@@ -146,7 +146,18 @@ class BlogPostTemplate extends React.Component {
                   flex: `0 0 auto`,
                 }}
               >
-                <Link to={post.frontmatter.author.fields.slug}>
+                <Link
+                  to={post.frontmatter.author.fields.slug}
+                  css={{
+                    "&&": {
+                      borderBottom: 0,
+                      boxShadow: `none`,
+                      "&:hover": {
+                        background: `none`,
+                      },
+                    },
+                  }}
+                >
                   <Img
                     fixed={post.frontmatter.author.avatar.childImageSharp.fixed}
                     css={{